fix(guardian-r1): resolve barrel export, cross-feature imports, raw HTML, and file-length violations
- Add useCreateProgressTask to tasks barrel export (Fix 1)
- Replace deep cross-feature imports with barrel imports in QuickCapture (Fix 2)
- Replace raw <button> with @ui Button (ghost variant) in QuickCapture (Fix 3)
- Replace raw <p> with @ui Text (variant=error) in MilestoneEditDialog (Fix 4)
- Replace raw <p> with @ui Text (variant=muted) in QuickCapture (Fix 5)
- Extract MilestoneTaskListEditor component to reduce MilestoneEditDialog from 379 to ~170 lines (Fix 6)
